Ensure message text is not cut off
When setting message text, make sure the minimal size is set accurately.
This ensures that even if Wrap() fails to do its job (which it may for
some languages that don't use spaces, e.g. Chinese, Japanese or Thai),
the text won't be cut off. The window may look a bit too wide and ugly,
but that's far better than not having the text shown fully.

Fixes #261.